moved cost responsibilty to booker, created abstract superclass Unavailable
