+++ lib/booker.rb
@@ -2,6 +2,7 @@ require_relative "../spec/spec_helper"
 
 module Hotel
   class Booker
+    SALES_TAX_ADJUSTMENT = 1.09
     attr_reader :manifest
 
     def initialize(manifest: Manifest.new)
@@ -11,6 +12,10 @@ module Hotel
     def book_room(reservation, room)
       room.unavailable << reservation
       return room
+    end
+
+    def total_cost_of_booking(reservation, room)
+      return room.cost_per_day * reservation.duration * SALES_TAX_ADJUSTMENT
     end
   end
 end
