+++ lib/reservation.rb
@@ -1,31 +1,10 @@
 require_relative "../spec/spec_helper"
+require_relative "unavailable"
 
 module Hotel
-  class Reservation
-    attr_reader :check_in, :check_out, :id
-    @@confirmation_number = 123000
-
+  class Reservation < Unavailable
     def initialize(check_in:, check_out:)
-      @check_in = Date.parse(check_in)
-      @check_out = Date.parse(check_out)
-      unless valid_date_range?(@check_in, @check_out)
-        raise InvalidDateRangeError.new()
-      end
-      @id = generate_confirmation_id
-    end
-
-    def valid_date_range?(check_in, check_out)
-      return check_in < check_out && check_in >= Time.new.to_date
-    end
-
-    private
-
-    def generate_confirmation_id(prefix = "R")
-      return "#{prefix}#{self.class.confirmation_number_generator}"
-    end
-
-    def self.confirmation_number_generator
-      @@confirmation_number += 1
+      super()
     end
   end
 end
