+++ spec/spec_helper.rb
@@ -15,3 +15,4 @@ require_relative "../lib/reservation.rb"
 require_relative "../lib/custom_exceptions.rb"
 require_relative "../lib/manifest.rb"
 require_relative "../lib/booker.rb"
+require_relative "../lib/unavailable.rb"
